feat: move custom list item component to its own file
